reorganize core a bit
