+++ src/actions.rs
@@ -1,502 +1,10 @@
-use std::ops::AddAssign;
-
-use crate::HashMap;
-use crate::{constraint::AllEqualTypeConstraint, typechecking::FuncType, *};
+use crate::core::{
+    CoreAction, CoreActions, ResolvedAtomTerm, ResolvedCoreActions, SpecializedPrimitive,
+};
+use crate::{typechecking::FuncType, *};
 use typechecking::TypeError;
 
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum HeadOrEq<Head> {
-    Symbol(Head),
-    Eq,
-}
-
-pub(crate) type SymbolOrEq = HeadOrEq<Symbol>;
-
-impl From<Symbol> for SymbolOrEq {
-    fn from(value: Symbol) -> Self {
-        SymbolOrEq::Symbol(value)
-    }
-}
-
-impl<Head> HeadOrEq<Head> {
-    pub fn is_eq(&self) -> bool {
-        matches!(self, HeadOrEq::Eq)
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct GenericCoreRule<BodyF, HeadF, Leaf> {
-    pub body: Query<BodyF, Leaf>,
-    pub head: CoreActions<HeadF, Leaf>,
-}
-
-pub(crate) type CoreRule = GenericCoreRule<SymbolOrEq, Symbol, Symbol>;
-pub(crate) type ResolvedCoreRule = GenericCoreRule<ResolvedCall, ResolvedCall, ResolvedVar>;
-
-#[derive(Debug, Clone)]
-pub(crate) struct SpecializedPrimitive {
-    pub(crate) primitive: Primitive,
-    pub(crate) input: Vec<ArcSort>,
-    pub(crate) output: ArcSort,
-}
-
-#[derive(Debug, Clone)]
-pub(crate) enum ResolvedCall {
-    Func(FuncType),
-    Primitive(SpecializedPrimitive),
-}
-
-impl SymbolLike for ResolvedCall {
-    fn to_symbol(&self) -> Symbol {
-        match self {
-            ResolvedCall::Func(f) => f.name,
-            ResolvedCall::Primitive(prim) => prim.primitive.0.name(),
-        }
-    }
-}
-
-impl ResolvedCall {
-    pub fn output(&self) -> &ArcSort {
-        match self {
-            ResolvedCall::Func(func) => &func.output,
-            ResolvedCall::Primitive(prim) => &prim.output,
-        }
-    }
-
-    // Different from `from_resolution`, this function only considers function types and not primitives.
-    // As a result, it only requires input argument types, so types.len() == func.input.len(),
-    // while for `from_resolution`, types.len() == func.input.len() + 1 to account for the output type
-    pub fn from_resolution_func_types(
-        head: &Symbol,
-        types: &[ArcSort],
-        typeinfo: &TypeInfo,
-    ) -> Option<ResolvedCall> {
-        if let Some(ty) = typeinfo.func_types.get(head) {
-            // As long as input types match, a result is returned.
-            let expected = ty.input.iter().map(|s| s.name());
-            let actual = types.iter().map(|s| s.name());
-            if expected.eq(actual) {
-                return Some(ResolvedCall::Func(ty.clone()));
-            }
-        }
-        None
-    }
-
-    pub fn from_resolution(head: &Symbol, types: &[ArcSort], typeinfo: &TypeInfo) -> ResolvedCall {
-        let mut resolved_call = Vec::with_capacity(1);
-        if let Some(ty) = typeinfo.func_types.get(head) {
-            let expected = ty.input.iter().chain(once(&ty.output)).map(|s| s.name());
-            let actual = types.iter().map(|s| s.name());
-            if expected.eq(actual) {
-                resolved_call.push(ResolvedCall::Func(ty.clone()));
-            }
-        }
-
-        if let Some(primitives) = typeinfo.primitives.get(head) {
-            for primitive in primitives {
-                if primitive.accept(types) {
-                    let (out, inp) = types.split_last().unwrap();
-                    resolved_call.push(ResolvedCall::Primitive(SpecializedPrimitive {
-                        primitive: primitive.clone(),
-                        input: inp.to_vec(),
-                        output: out.clone(),
-                    }));
-                }
-            }
-        }
-
-        assert!(resolved_call.len() == 1);
-        resolved_call.pop().unwrap()
-    }
-}
-
-impl Display for ResolvedCall {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            ResolvedCall::Func(func) => write!(f, "{}", func.name),
-            ResolvedCall::Primitive(prim) => write!(f, "{}", prim.primitive.0.name()),
-        }
-    }
-}
-
-impl ToSexp for ResolvedCall {
-    fn to_sexp(&self) -> Sexp {
-        Sexp::Symbol(self.to_string())
-    }
-}
-
-impl<Leaf: Clone> Query<ResolvedCall, Leaf> {
-    pub fn filters(&self) -> impl Iterator<Item = GenericAtom<Primitive, Leaf>> + '_ {
-        self.atoms.iter().filter_map(|atom| match &atom.head {
-            ResolvedCall::Func(_) => None,
-            ResolvedCall::Primitive(head) => Some(GenericAtom {
-                head: head.primitive.clone(),
-                args: atom.args.clone(),
-            }),
-        })
-    }
-
-    pub fn funcs(&self) -> impl Iterator<Item = GenericAtom<Symbol, Leaf>> + '_ {
-        self.atoms.iter().filter_map(|atom| match &atom.head {
-            ResolvedCall::Func(head) => Some(GenericAtom {
-                head: head.name,
-                args: atom.args.clone(),
-            }),
-            ResolvedCall::Primitive(_) => None,
-        })
-    }
-}
-
-impl<Head1, Head2, Leaf> GenericCoreRule<Head1, Head2, Leaf>
-where
-    Head1: Clone,
-    Head2: Clone,
-    Leaf: Clone + Eq + Hash,
-{
-    pub fn subst(&mut self, subst: &HashMap<Leaf, GenericAtomTerm<Leaf>>) {
-        for atom in &mut self.body.atoms {
-            atom.subst(subst);
-        }
-        self.head.subst(subst);
-    }
-}
-
-impl<Head, Leaf> GenericRule<Head, Leaf, ()>
-where
-    Leaf: Clone + Eq + Hash + Debug,
-    Head: Clone,
-{
-    pub(crate) fn to_core_rule(
-        &self,
-        typeinfo: &TypeInfo,
-        mut fresh_gen: impl FreshGen<Head, Leaf>,
-    ) -> Result<GenericCoreRule<HeadOrEq<Head>, Head, Leaf>, TypeError>
-    where
-        Leaf: SymbolLike,
-    {
-        let GenericRule { head, body } = self;
-
-        let (body, _correspondence) = Facts(body.clone()).to_query(typeinfo, &mut fresh_gen);
-        let mut binding = body.get_vars();
-        let (head, _correspondence) =
-            head.to_core_actions(typeinfo, &mut binding, &mut fresh_gen)?;
-        Ok(GenericCoreRule { body, head })
-    }
-
-    fn to_canonicalized_core_rule_impl(
-        &self,
-        typeinfo: &TypeInfo,
-        fresh_gen: impl FreshGen<Head, Leaf>,
-        value_eq: impl Fn(&GenericAtomTerm<Leaf>, &GenericAtomTerm<Leaf>) -> Head,
-    ) -> Result<GenericCoreRule<Head, Head, Leaf>, TypeError>
-    where
-        Leaf: SymbolLike,
-    {
-        let rule = self.to_core_rule(typeinfo, fresh_gen)?;
-        let rule = rule.canonicalize(value_eq);
-        Ok(rule)
-    }
-}
-
-impl ResolvedRule {
-    pub(crate) fn to_canonicalized_core_rule(
-        &self,
-        typeinfo: &TypeInfo,
-    ) -> Result<ResolvedCoreRule, TypeError> {
-        let value_eq = &typeinfo.primitives.get(&Symbol::from("value-eq")).unwrap()[0];
-        let unit = typeinfo.get_sort_nofail::<UnitSort>();
-        self.to_canonicalized_core_rule_impl(typeinfo, ResolvedGen::new(), |at1, at2| {
-            ResolvedCall::Primitive(SpecializedPrimitive {
-                primitive: value_eq.clone(),
-                input: vec![at1.output(typeinfo), at2.output(typeinfo)],
-                output: unit.clone(),
-            })
-        })
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub enum GenericAtomTerm<Leaf> {
-    Var(Leaf),
-    Literal(Literal),
-    Global(Leaf),
-}
-
-pub type AtomTerm = GenericAtomTerm<Symbol>;
-pub type ResolvedAtomTerm = GenericAtomTerm<ResolvedVar>;
-
-impl AtomTerm {
-    pub fn to_expr(&self) -> Expr {
-        match self {
-            AtomTerm::Var(v) => Expr::Var((), *v),
-            AtomTerm::Literal(l) => Expr::Lit((), l.clone()),
-            AtomTerm::Global(v) => Expr::Var((), *v),
-        }
-    }
-}
-
-impl ResolvedAtomTerm {
-    pub fn output(&self, typeinfo: &TypeInfo) -> ArcSort {
-        match self {
-            ResolvedAtomTerm::Var(v) => v.sort.clone(),
-            ResolvedAtomTerm::Literal(l) => typeinfo.infer_literal(l),
-            ResolvedAtomTerm::Global(v) => v.sort.clone(),
-        }
-    }
-}
-
-impl std::fmt::Display for AtomTerm {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            AtomTerm::Var(v) => write!(f, "{}", v),
-            AtomTerm::Literal(lit) => write!(f, "{}", lit),
-            AtomTerm::Global(g) => write!(f, "{}", g),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-pub struct GenericAtom<Head, Leaf> {
-    pub head: Head,
-    pub args: Vec<GenericAtomTerm<Leaf>>,
-}
-
-pub type Atom<T> = GenericAtom<T, Symbol>;
-
-impl<T: std::fmt::Display> std::fmt::Display for Atom<T> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "({} {}) ", self.head, ListDisplay(&self.args, " "))
-    }
-}
-
-#[derive(Debug, Clone)]
-pub struct Query<Head, Leaf> {
-    pub atoms: Vec<GenericAtom<Head, Leaf>>,
-}
-
-impl<Head, Leaf> Default for Query<Head, Leaf> {
-    fn default() -> Self {
-        Self {
-            atoms: Default::default(),
-        }
-    }
-}
-
-impl Query<SymbolOrEq, Symbol> {
-    pub fn get_constraints(
-        &self,
-        type_info: &TypeInfo,
-    ) -> Result<Vec<Constraint<AtomTerm, ArcSort>>, TypeError> {
-        let mut constraints = vec![];
-        for atom in self.atoms.iter() {
-            constraints.extend(atom.get_constraints(type_info)?.into_iter());
-        }
-        Ok(constraints)
-    }
-
-    pub(crate) fn atom_terms(&self) -> HashSet<AtomTerm> {
-        self.atoms
-            .iter()
-            .flat_map(|atom| atom.args.iter().cloned())
-            .collect()
-    }
-}
-
-impl<Head, Leaf> Query<Head, Leaf>
-where
-    Leaf: Eq + Clone + Hash,
-    Head: Clone,
-{
-    pub(crate) fn get_vars(&self) -> IndexSet<Leaf> {
-        self.atoms
-            .iter()
-            .flat_map(|atom| atom.vars())
-            .collect::<IndexSet<_>>()
-    }
-}
-
-impl<Head, Leaf> AddAssign for Query<Head, Leaf> {
-    fn add_assign(&mut self, rhs: Self) {
-        self.atoms.extend(rhs.atoms);
-    }
-}
-
-impl std::fmt::Display for Query<Symbol, Symbol> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for atom in &self.atoms {
-            writeln!(f, "{atom}")?;
-        }
-        Ok(())
-    }
-}
-
-impl std::fmt::Display for Query<ResolvedCall, Symbol> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        for atom in self.funcs() {
-            writeln!(f, "{atom}")?;
-        }
-        let filters: Vec<_> = self.filters().collect();
-        if !filters.is_empty() {
-            writeln!(f, "where ")?;
-            for filter in &filters {
-                writeln!(
-                    f,
-                    "({} {})",
-                    filter.head.name(),
-                    ListDisplay(&filter.args, " ")
-                )?;
-            }
-        }
-        Ok(())
-    }
-}
-
-impl<Head, Leaf> GenericAtom<Head, Leaf>
-where
-    Leaf: Clone + Eq + Hash,
-    Head: Clone,
-{
-    pub fn vars(&self) -> impl Iterator<Item = Leaf> + '_ {
-        self.args.iter().filter_map(|t| match t {
-            GenericAtomTerm::Var(v) => Some(v.clone()),
-            GenericAtomTerm::Literal(_) => None,
-            GenericAtomTerm::Global(_) => None,
-        })
-    }
-
-    fn subst(&mut self, subst: &HashMap<Leaf, GenericAtomTerm<Leaf>>) {
-        for arg in self.args.iter_mut() {
-            match arg {
-                GenericAtomTerm::Var(v) => {
-                    if let Some(at) = subst.get(v) {
-                        *arg = at.clone();
-                    }
-                }
-                GenericAtomTerm::Literal(_) => (),
-                GenericAtomTerm::Global(_) => (),
-            }
-        }
-    }
-}
-impl Atom<Symbol> {
-    pub(crate) fn to_expr(&self) -> Expr {
-        let n = self.args.len();
-        Expr::Call(
-            (),
-            self.head,
-            self.args[0..n - 1]
-                .iter()
-                .map(|arg| arg.to_expr())
-                .collect(),
-        )
-    }
-}
-
-pub(crate) struct ValueEq {
-    pub unit: Arc<UnitSort>,
-}
-
-impl PrimitiveLike for ValueEq {
-    fn name(&self) -> Symbol {
-        "value-eq".into()
-    }
-
-    fn get_type_constraints(&self) -> Box<dyn TypeConstraint> {
-        AllEqualTypeConstraint::new(self.name())
-            .with_exact_length(3)
-            .with_output_sort(self.unit.clone())
-            .into_box()
-    }
-
-    fn apply(&self, values: &[Value], _egraph: &EGraph) -> Option<Value> {
-        assert_eq!(values.len(), 2);
-        if values[0] == values[1] {
-            Some(Value::unit())
-        } else {
-            None
-        }
-    }
-}
-
-impl<Head, Leaf> GenericCoreRule<HeadOrEq<Head>, Head, Leaf>
-where
-    Leaf: Eq + Clone + Hash + Debug,
-    Head: Clone,
-{
-    /// Transformed a UnresolvedCoreRule into a CanonicalizedCoreRule.
-    /// In particular, it removes equality checks between variables and
-    /// other arguments, and turns equality checks between non-variable arguments
-    /// into a primitive equality check `value-eq`.
-    pub(crate) fn canonicalize(
-        self,
-        // Users need to pass in a substitute for equality constraints.
-        value_eq: impl Fn(&GenericAtomTerm<Leaf>, &GenericAtomTerm<Leaf>) -> Head,
-    ) -> GenericCoreRule<Head, Head, Leaf> {
-        let mut result_rule = self;
-        loop {
-            let mut to_subst = None;
-            for atom in result_rule.body.atoms.iter() {
-                if atom.head.is_eq() && atom.args[0] != atom.args[1] {
-                    match &atom.args[..] {
-                        [GenericAtomTerm::Var(x), y] | [y, GenericAtomTerm::Var(x)] => {
-                            to_subst = Some((x, y));
-                            break;
-                        }
-                        _ => (),
-                    }
-                }
-            }
-            if let Some((x, y)) = to_subst {
-                let subst = HashMap::from_iter(vec![(x.clone(), y.clone())]);
-                result_rule.subst(&subst);
-            } else {
-                break;
-            }
-        }
-
-        let atoms = result_rule
-            .body
-            .atoms
-            .into_iter()
-            .filter_map(|atom| match atom.head {
-                HeadOrEq::Eq => {
-                    assert_eq!(atom.args.len(), 2);
-                    match (&atom.args[0], &atom.args[1]) {
-                        (GenericAtomTerm::Var(v1), GenericAtomTerm::Var(v2)) => {
-                            assert_eq!(v1, v2);
-                            None
-                        }
-                        (GenericAtomTerm::Var(_), _) | (_, GenericAtomTerm::Var(_)) => {
-                            panic!("equalities between variable and non-variable arguments should have been canonicalized")
-                        }
-                        (at1, at2) => {
-                            if at1 == at2 {
-                                None
-                            } else {
-                                Some(GenericAtom {
-                                    head: value_eq(&atom.args[0], &atom.args[1]),
-                                    args: vec![
-                                        atom.args[0].clone(),
-                                        atom.args[1].clone(),
-                                        GenericAtomTerm::Literal(Literal::Unit),
-                                    ],
-                                })
-                            }
-                        },
-                    }
-                }
-                HeadOrEq::Symbol(symbol) => Some(GenericAtom {
-                    head: symbol,
-                    args: atom.args,
-                }),
-            })
-            .collect();
-        GenericCoreRule {
-            body: Query { atoms },
-            head: result_rule.head,
-        }
-    }
-}
+use crate::{ast::Literal, core::ResolvedCall, ExtractReport, Value};
 
 struct ActionCompiler<'a> {
     egraph: &'a EGraph,
@@ -506,7 +14,7 @@ struct ActionCompiler<'a> {
 }
 
 impl<'a> ActionCompiler<'a> {
-    fn check_action(&mut self, action: &CoreAction<ResolvedCall, ResolvedVar>) {
+    fn compile_action(&mut self, action: &CoreAction<ResolvedCall, ResolvedVar>) {
         match action {
             CoreAction::Let(v, f, args) => {
                 self.do_call(f, args);
@@ -638,7 +146,7 @@ impl EGraph {
         for var in binding {
             types.insert(var.name, var.sort.clone());
         }
-        let mut checker = ActionCompiler {
+        let mut compiler = ActionCompiler {
             egraph: self,
             types: &types,
             locals: IndexSet::default(),
@@ -646,13 +154,13 @@ impl EGraph {
         };
 
         for a in actions.0.iter() {
-            checker.check_action(a);
+            compiler.compile_action(a);
         }
 
-        Ok(Program(checker.instructions))
+        Ok(Program(compiler.instructions))
     }
 
-    // This is the ugly part of the code. CoreActions lowered from
+    // This is the ugly part. CoreActions lowered from
     // expressions like `2` is an empty vector, because no action is taken.
     // So to explicitly obtain the return value of an expression, compile_expr
     // needs to also take a `target`.`
@@ -666,7 +174,7 @@ impl EGraph {
         for var in binding {
             types.insert(var.name, var.sort.clone());
         }
-        let mut checker = ActionCompiler {
+        let mut compiler = ActionCompiler {
             egraph: self,
             types: &types,
             locals: IndexSet::default(),
@@ -674,11 +182,11 @@ impl EGraph {
         };
 
         for a in actions.0.iter() {
-            checker.check_action(a);
+            compiler.compile_action(a);
         }
-        checker.do_atom_term(target);
+        compiler.do_atom_term(target);
 
-        Ok(Program(checker.instructions))
+        Ok(Program(compiler.instructions))
     }
 
     fn perform_set(
