+++ src/constraint.rs
@@ -3,8 +3,8 @@ use crate::{
         CoreActions, GenericAction, GenericActions, GenericExpr, GenericFact, MappedAction,
         NormAction, ResolvedAction, ResolvedActions, ResolvedExpr, ResolvedFact, ResolvedVar,
     },
+    core::{Atom, AtomTerm, CoreRule, Query, ResolvedCall, SymbolOrEq},
     sort::I64Sort,
-    typecheck::{Atom, AtomTerm, CoreRule, Query, ResolvedCall, SymbolOrEq},
     typechecking::TypeError,
     util::{FreshGen, HashMap, HashSet, SymbolGen},
     ArcSort, Symbol, TypeInfo,
